feat: rename interceptor file
Ref: #363
